chore(renovate): refactor code to address new lint rules (react-refresh and deprecated warnings)
