+++ hivemq-edge/src/frontend/src/__test-utils__/hooks/WrapperEdgeProvider.tsx
@@ -1,7 +1,8 @@
 import type { FC, PropsWithChildren } from 'react'
-import { EdgeFlowProvider } from '@/modules/Workspace/hooks/FlowContext.tsx'
-import type { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
+
 import { SimpleWrapper } from '@/__test-utils__/hooks/SimpleWrapper.tsx'
+import type { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
+import { EdgeFlowProvider } from '@/modules/Workspace/hooks/EdgeFlowProvider.tsx'
 
 export const getWrapperEdgeProvider = (defaults?: Partial<EdgeFlowOptions>) => {
   const WrapperEdgeProvider: FC<PropsWithChildren> = ({ children }) => {
