+++ hivemq-edge/src/frontend/src/__test-utils__/react-flow/providers.tsx
@@ -1,7 +1,8 @@
 import type { ReactNode } from 'react'
-import { EdgeFlowProvider } from '@/modules/Workspace/hooks/FlowContext.tsx'
 import { ReactFlowProvider } from 'reactflow'
 import type { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
+import { EdgeFlowProvider } from '@/modules/Workspace/hooks/EdgeFlowProvider.tsx'
+
 import 'reactflow/dist/style.css'
 
 export const mockReactFlow = (children: ReactNode, defaults?: Partial<EdgeFlowOptions>) => (
