+++ hivemq-edge/src/frontend/src/components/MQTT/types.ts
@@ -1,0 +1,5 @@
+export enum SelectEntityType {
+  TOPIC = 'TOPIC',
+  TOPIC_FILTER = 'TOPIC_FILTER',
+  TAG = 'TAG',
+}
