+++ hivemq-edge/src/frontend/src/components/PaginatedTable/components/PaginationBar.spec.cy.tsx
@@ -9,7 +9,6 @@ const MOCK_PAGE_SIZES = [5, 10, 20, 30, 40, 50]
 const MOCK_PAGE_COUNT = 10
 const MOCK_PAGE_INDEX = 2
 
-/* eslint-disable @typescript-eslint/no-empty-function */
 const MOCK_TABLE: Partial<Table<Adapter>> = {
   initialState: undefined,
   setPageIndex: () => {},
