+++ hivemq-edge/src/frontend/src/components/rjsf/MqttTransformation/types.ts
@@ -1,0 +1,5 @@
+export enum MappingStrategy {
+  EXACT = 'EXACT',
+  TYPED = 'TYPED',
+  TRANSFORMED = 'TRANSFORMED',
+}
