+++ hivemq-edge/src/frontend/src/extensions/datahub/components/controls/DesignerToolbox.spec.cy.tsx
@@ -26,7 +26,6 @@ describe('DesignerToolbox', () => {
     cy.mountWithProviders(<DesignerToolbox />, { wrapper })
     cy.getByTestId('toolbox-trigger').click()
 
-    // eslint-disable-next-line cypress/no-unnecessary-waiting
     cy.wait(100) // Wait for dropdown (ugly)
 
     cy.checkAccessibility()
