+++ hivemq-edge/src/frontend/src/extensions/datahub/routes.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-refresh/only-export-components */
 import { lazy } from 'react'
 import type { RouteObject } from 'react-router-dom'
