+++ hivemq-edge/src/frontend/src/modules/App/routes.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-refresh/only-export-components */
 import { lazy } from 'react'
 import { createBrowserRouter } from 'react-router-dom'
