+++ hivemq-edge/src/frontend/src/modules/Auth/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import { AuthContext } from '../AuthProvider.tsx'
+import { AuthContext } from '@/modules/Auth/AuthContext.ts'
 
 export const useAuth = () => {
   const context = useContext(AuthContext)
