+++ hivemq-edge/src/frontend/src/modules/Bridges/hooks/useBridgeConfig.tsx
@@ -1,16 +1,9 @@
+import { useContext } from 'react'
 import type { Bridge } from '@/api/__generated__'
-import type { Dispatch, FunctionComponent, PropsWithChildren, SetStateAction } from 'react'
-import { createContext, useContext, useState } from 'react'
-
-interface BridgeContextProps {
-  bridge: Bridge
-  setBridge: Dispatch<SetStateAction<Bridge>>
-}
-
-export const BridgeContext = createContext<BridgeContextProps | null>(null)
+import { BridgeContext } from './BridgeContext'
 
 // TODO The number and booleans should all be coming from the openAPI specs since they are all mandatory (see $Bridge.properties.cleanStart.isRequired)
-// eslint-disable-next-line react-refresh/only-export-components
+
 export const bridgeInitialState: Bridge = {
   cleanStart: true,
   host: '',
@@ -21,13 +14,6 @@ export const bridgeInitialState: Bridge = {
   persist: true,
 }
 
-export const BridgeProvider: FunctionComponent<PropsWithChildren> = ({ children }) => {
-  const [bridge, setBridge] = useState<Bridge>(bridgeInitialState)
-
-  return <BridgeContext.Provider value={{ bridge, setBridge }}>{children}</BridgeContext.Provider>
-}
-
-// eslint-disable-next-line react-refresh/only-export-components
 export const useBridgeSetup = () => {
   const context = useContext(BridgeContext)
   if (context === null) {
