+++ hivemq-edge/src/frontend/src/modules/Workspace/EdgeFlowPage.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'react-i18next'
 
 import PageContainer from '@/components/PageContainer.tsx'
 
-import { EdgeFlowProvider } from './hooks/FlowContext.tsx'
-import ReactFlowWrapper from './components/ReactFlowWrapper.tsx'
-import WorkspaceOptionsDrawer from './components/drawers/WorkspaceOptionsDrawer.tsx'
+import { EdgeFlowProvider } from '@/modules/Workspace/hooks/EdgeFlowProvider.tsx'
+import ReactFlowWrapper from '@/modules/Workspace/components/ReactFlowWrapper.tsx'
+import WorkspaceOptionsDrawer from '@/modules/Workspace/components/drawers/WorkspaceOptionsDrawer.tsx'
 
 const EdgeFlowPage: FC = () => {
   const { t } = useTranslation()
