+++ hivemq-edge/src/frontend/src/modules/Workspace/hooks/EdgeFlowProvider.spec.tsx
@@ -1,8 +1,9 @@
 import { useContext } from 'react'
 import { describe, it, expect } from 'vitest'
 import { render, act } from '@testing-library/react'
-import { EdgeFlowContext, EdgeFlowProvider } from './FlowContext.tsx'
+import { EdgeFlowContext } from './FlowContext.tsx'
 import type { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
+import { EdgeFlowProvider } from '@/modules/Workspace/hooks/EdgeFlowProvider.tsx'
 
 const optionKeys = ['showTopics', 'showStatus', 'showMetrics', 'showGateway']
