+++ hivemq-edge/src/frontend/src/modules/Workspace/hooks/FlowContext.tsx
@@ -1,10 +1,8 @@
-import type { Dispatch, FunctionComponent, PropsWithChildren, SetStateAction } from 'react'
-import { createContext, useState } from 'react'
+import type { Dispatch, SetStateAction } from 'react'
+import { createContext } from 'react'
 import type { UseDisclosureReturn } from '@chakra-ui/react'
-import { useDisclosure } from '@chakra-ui/react'
 
 import type { EdgeFlowGrouping, EdgeFlowOptions } from '../types.ts'
-import { EdgeFlowLayout } from '../types.ts'
 
 export interface EdgeFlowContextType {
   options: EdgeFlowOptions
@@ -14,31 +12,4 @@ export interface EdgeFlowContextType {
   optionDrawer: UseDisclosureReturn
 }
 
-const defaultEdgeFlowContext: EdgeFlowOptions = {
-  showTopics: true,
-  showStatus: true,
-  showGateway: false,
-}
-
-const defaultEdgeFlowGrouping: EdgeFlowGrouping = {
-  keys: [],
-  showGroups: false,
-  layout: EdgeFlowLayout.HORIZONTAL,
-}
-
 export const EdgeFlowContext = createContext<EdgeFlowContextType | null>(null)
-
-export const EdgeFlowProvider: FunctionComponent<PropsWithChildren<{ defaults?: Partial<EdgeFlowOptions> }>> = ({
-  children,
-  defaults,
-}) => {
-  const [options, setOptions] = useState<EdgeFlowOptions>({ ...defaultEdgeFlowContext, ...defaults })
-  const [groups, setGroups] = useState<EdgeFlowGrouping>(defaultEdgeFlowGrouping)
-  const optionDrawer = useDisclosure()
-
-  return (
-    <EdgeFlowContext.Provider value={{ options, setOptions, groups, setGroups, optionDrawer }}>
-      {children}
-    </EdgeFlowContext.Provider>
-  )
-}
